% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_book.R
\name{download_book}
\alias{download_book}
\title{download_book}
\usage{
download_book(save_loc = rstudioapi::selectDirectory())
}
\arguments{
\item{save_loc}{\strong{Run with no arguments to pick a directory interactively} Directory the book should be downloaded to}
}
\description{
Need to refer to the Epidemiology R Handbook offline in the field?
Easily download the handbook in advance with this function
}
