% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handbookfiles-package.R
\docType{package}
\name{handbookfiles-package}
\alias{handbookfiles}
\alias{handbookfiles-package}
\title{handbookfiles: Data Used in The Epidemiology R Handbook}
\description{
Data used in the Epidemiologist R Handbook, 
  an open-access R reference manual for applied epidemiologists and public health practitioners.
}
\author{
\strong{Maintainer}: Isaac Florence \email{isaac.florence@outlook.com} (\href{https://orcid.org/0000-0001-8260-1237}{ORCID})

}
\keyword{internal}
