% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio_import.R
\name{import_erh}
\alias{import_erh}
\title{import_erh

Epi R Handbook data drop-in replacement for rio::import, as used in the handbook}
\usage{
import_erh(path = NULL)
}
\arguments{
\item{data}{}
}
\description{
import_erh

Epi R Handbook data drop-in replacement for rio::import, as used in the handbook
}
